+++ OAR/headers/terrain/TerrainChunk.h
@@ -10,7 +10,7 @@ namespace ORNG {
 		friend class TerrainQuadtree;
 		friend class ChunkLoader;
 	public:
-		TerrainChunk(glm::vec3 bot_left_coord, float resolution, unsigned int width, unsigned int seed, float height_scale) :
+		TerrainChunk(glm::vec3 bot_left_coord, unsigned int resolution, unsigned int width, unsigned int seed, float height_scale) :
 			m_bot_left_coord(bot_left_coord), m_resolution(resolution), m_width(width), m_seed(seed), m_height_scale(height_scale) {};
 
 		void LoadGLData();
@@ -23,7 +23,7 @@ namespace ORNG {
 		TerrainGenerator::TerrainData data;
 
 		glm::vec3 m_bot_left_coord;
-		float m_resolution;
+		unsigned int m_resolution;
 		float m_height_scale;
 		unsigned int m_width;
 		unsigned int m_seed;
